centralize site and nwo logic with RepositoryFinder
