+++ .rubocop.yml
@@ -11,3 +11,7 @@ Metrics/LineLength:
 Metrics/BlockLength:
   Exclude:
     - spec/**/*
+    
+Metrics/MethodLength:
+    Exclude:
+      - spec/**/
