+++ spec/spec_helper.rb
@@ -1,6 +1,7 @@
 require "jekyll-github-metadata"
 require "webmock/rspec"
 require "pathname"
+require "jekyll"
 
 SPEC_DIR = Pathname.new(File.expand_path("../", __FILE__))
 
@@ -110,6 +111,44 @@ module EnvHelper
     end
     raise ArgumentError, "Expect 2 strings or a Hash of VAR => VAL"
   end
+end
+
+def expected_values
+  {
+    "environment"          => "dotcom",
+    "hostname"             => "github.com",
+    "pages_env"            => "dotcom",
+    "pages_hostname"       => "github.io",
+    "help_url"             => "https://help.github.com",
+    "api_url"              => "https://api.github.com",
+    "versions"             => {},
+    "public_repositories"  => Regexp.new('"id"=>17261694, "name"=>"atom-jekyll"'),
+    "organization_members" => Regexp.new('"login"=>"parkr", "id"=>237985'),
+    "build_revision"       => %r![a-f0-9]{40}!,
+    "project_title"        => "github-metadata",
+    "project_tagline"      => ":octocat: `site.github`",
+    "owner_name"           => "jekyll",
+    "owner_url"            => "https://github.com/jekyll",
+    "owner_gravatar_url"   => "https://github.com/jekyll.png",
+    "repository_url"       => "https://github.com/jekyll/github-metadata",
+    "repository_nwo"       => "jekyll/github-metadata",
+    "repository_name"      => "github-metadata",
+    "zip_url"              => "https://github.com/jekyll/github-metadata/zipball/gh-pages",
+    "tar_url"              => "https://github.com/jekyll/github-metadata/tarball/gh-pages",
+    "clone_url"            => "https://github.com/jekyll/github-metadata.git",
+    "releases_url"         => "https://github.com/jekyll/github-metadata/releases",
+    "issues_url"           => "https://github.com/jekyll/github-metadata/issues",
+    "wiki_url"             => nil, # disabled
+    "language"             => "Ruby",
+    "is_user_page"         => false,
+    "is_project_page"      => true,
+    "show_downloads"       => true,
+    "url"                  => "http://jekyll.github.io/github-metadata",
+    "baseurl"              => "/github-metadata",
+    "contributors"         => %r!"login"=>"parkr", "id"=>237985!,
+    "releases"             => %r!"tag_name"=>"v1.1.0"!,
+    "latest_release"       => %r!assets_url!,
+  }
 end
 
 RSpec.configure do |config|
